rename code-gen to filegen since that's more apt description
