+++ pkg/code-generator/filegen/templates/event_loop_template.go
