+++ pkg/code-generator/filegen/templates/event_loop_test_template.go
