+++ pkg/code-generator/filegen/templates/resource_client_template.go
