+++ pkg/code-generator/filegen/templates/resource_client_test_template.go
