+++ pkg/code-generator/filegen/templates/resource_reconciler_template.go
