+++ pkg/code-generator/filegen/templates/resource_template.go
