+++ pkg/code-generator/filegen/templates/simple_event_loop_template.go
