+++ pkg/code-generator/filegen/templates/snapshot_emitter_template.go
