+++ pkg/code-generator/filegen/templates/snapshot_emitter_test_template.go
