+++ pkg/code-generator/filegen/templates/snapshot_simple_emitter_template.go
