+++ pkg/code-generator/filegen/templates/snapshot_template.go
