+++ pkg/code-generator/filegen/templates/xds_template.go
