+++ pkg/code-generator/filegen/utils/is_project_resource.go
